# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/integrations": {
      "get": {
        "operationId": "IntegrationsController_listIntegrations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "workspaceId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "additionalProperties": true
                          }
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "nullable": true,
                          "example": "req_123"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "requestId",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "List integrations in the current workspace",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/integrations/{integrationId}": {
      "patch": {
        "operationId": "IntegrationsController_updateIntegration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "connected",
                      "warning",
                      "error",
                      "syncing"
                    ],
                    "example": "connected"
                  },
                  "configJson": {
                    "type": "object",
                    "additionalProperties": true,
                    "example": {
                      "syncWindowDays": 30
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Update integration status/configuration",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/api/v1/recommendations": {
      "get": {
        "operationId": "RecommendationsController_listRecommendations",
        "parameters": [
          {
            "name": "domain",
            "required": false,
            "in": "query",
            "schema": {
              "example": "technical"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "new"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "createdAt:desc"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "example": 25
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "workspaceId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "additionalProperties": true
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "nullable": true,
                          "example": "req_123"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "requestId",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "List recommendations with pagination/filter/sort",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/recommendations/generate": {
      "post": {
        "operationId": "RecommendationsController_generateRecommendation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Reduce LCP on category pages"
                  },
                  "category": {
                    "type": "string",
                    "example": "technical"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Generate a recommendation for workspace",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/recommendations/{recommendationId}/approve": {
      "patch": {
        "operationId": "RecommendationsController_approveRecommendation",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Approve a recommendation",
        "tags": [
          "recommendations"
        ]
      }
    },
...
  },
  "info": {
    "title": "WebOps Wizard API",
    "description": "REST API for WebOps Wizard. All responses are wrapped in success/error envelopes under /api/v1.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wow_access",
        "description": "Session access token cookie"
      }
    },
...
  }
}